+++ client/src/components/KanbanBoard.tsx
@@ -434,7 +434,6 @@ export function KanbanBoard({
                     >
                       <KanbanColumnContent
                         stageId={stage.id}
-                        stageName={stage.name}
                         stageTasks={stageTasks}
                         allSubStages={allSubStages}
                         stageColor={stageColor}
@@ -471,7 +470,6 @@ export function KanbanBoard({
                   >
                     <KanbanColumnContent
                       stageId={stage.id}
-                      stageName={stage.name}
                       stageTasks={stageTasks}
                       allSubStages={allSubStages}
                       stageColor={stageColor}
@@ -558,7 +556,6 @@ export function KanbanBoard({
           activeId={activeId}
           activeTasks={activeTasks}
           stageColorMap={stageColorMap}
-          sortedStages={sortedStages}
           viewMode={viewMode}
         />
       </TaskSelectionContext.Provider>